AI_sales_bot/requirements.txt
# Core dependencies for chatbot and PDF maker
python-dotenv>=1.0.0
openai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pymongo>=4.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
pydantic>=2.0.0
streamlit>=1.28.0
requests>=2.31.0

# Financial analysis dependencies
yfinance>=0.2.0
pandas>=2.0.0
plotly>=5.0.0
